Collect FieldValue classes in same module
